project21_credit_scoring/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3
lightgbm==4.1.0

# Data Processing
imbalanced-learn==0.11.0
category-encoders==2.6.2

# Explainability
shap==0.43.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# API
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Web UI
streamlit==1.29.0
gradio==4.8.0

# Monitoring
prometheus-client==0.19.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
loguru==0.7.2

# Data Generation (for demo)
faker==20.1.0
